Make vesion resolution tests more realistic
The first iteration of these tests provided a set of constraints in
a single project: this made it easy to resolve, as we had _all_ of the
constraints available when resolved any of them to a version.

The test has been updated to use a chain of project dependencies to
bring in the constraints, forcing them to be resolved sequentially:
this exposes a couple of cases that are not supported by the current
implementation.